requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9
aiofiles==0.24.0
jinja2==3.1.2
email-validator==2.1.0